+++ vite.site.config.ts
@@ -10,6 +10,7 @@ export default defineConfig({
     rollupOptions: {
       input: {
         main: new URL("./site/index.html", import.meta.url).pathname,
+        notFound: new URL("./site/404.html", import.meta.url).pathname,
         privacy: new URL("./site/privacy/index.html", import.meta.url).pathname,
         terms: new URL("./site/terms/index.html", import.meta.url).pathname,
       },